+++ src/reportlab/rl_settings.py
@@ -70,7 +70,8 @@ trustedSchemes
 renderPMBackend
 xmlParser
 textPaths
-toColorCanUse'''.split())
+toColorCanUse
+defCWRF'''.split())
 
 allowTableBoundsErrors =    1 # set to 0 to die on too large elements in tables in debug (recommend 1 for production use)
 shapeChecking =             1
@@ -165,6 +166,8 @@ textPaths='freetype'                                #freetype or _renderPM or ba
                                                     #determines what code is used to create Paths from str
                                                     #see reportlab/graphics/utils.py for full horror
 toColorCanUse='rl_extended_literal_eval'            #change to None or 'rl_safe_eval' depending on trust
+defCWRF=0.02                                        #fraction we can reduce defined column widths for overcommitted
+                                                    #undefined widths
 
 # places to look for T1Font information
 T1SearchPath =  (