+++ CHANGES.md
@@ -11,9 +11,12 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
-CHANGES  4.0.8   ??/??/2023
+CHANGES  4.0.8   11/12/2023
 ---------------------------
     * fix rparsexml bug contributed by Matt Folwell  mjf at pearson co uk 
+    * allow class ddf attributes for Labels
+    * add rl_settings.defCWRF defined col width reduction fraction
+    * fix table wrapping bug contributed by Manuel Koch manuel koch at shinefour de
 
 CHANGES  4.0.7   06/11/2023
 ---------------------------